backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-compress==1.15
python-dotenv==1.0.1
bcrypt==4.0.1
reportlab==4.0.7
